pyproject.toml
[project]
name = "github-agent"
version = "0.1.0"
description = "GitHub PR Review Agent"
dependencies = [
    "git-filter-repo>=2.47.0",
    "portia-sdk-python",
    "python-dotenv",
    "requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
